Move GH16805 test into ntuple_processor.cxx
