+++ tree/ntuple/test/CMakeLists.txt
@@ -187,5 +187,3 @@ if(pyroot)
   ROOT_ADD_PYUNITTEST(ntuple_py_model ntuple_model.py)
 
 endif()
-
-ROOT_ADD_GTEST(ntuple_processor_gh16805 ntuple_processor_gh16805.cxx LIBRARIES ROOTNTuple)
